refactor(data-model): move the codec registry to `codec-common`
`CodecRegistry` and `createDefaultRegistry()` index and build the set of
`FabricCodec`s, which is wire-format-agnostic machinery; move them from
`codec-json` to `codec-common` alongside the codecs they hold, together
with `test/codec-json/CodecRegistry.test.ts`.

`codec-json` leaves nothing behind: its barrel drops both exports,
`codec-common`'s barrel picks them up, and `JsonEncodingContext` imports
them from the new location.
