+++ docs/plans/first-class-serializable-factories.md
@@ -195,10 +195,10 @@ Expected implementation files:
 - [ ] Decode to a frozen branded callable shell whose body throws
   `factory requires runner materialization` and whose state can be re-encoded.
 - [ ] Add a dedicated callable-factory codec slot to
-  `packages/data-model/src/codec-json/CodecRegistry.ts`; do not classify all
+  `packages/data-model/src/codec-common/CodecRegistry.ts`; do not classify all
   functions as primitives or match `Function` by constructor.
 - [ ] Register the codec in
-  `packages/data-model/src/codec-json/createDefaultRegistry.ts`.
+  `packages/data-model/src/codec-common/createDefaultRegistry.ts`.
 - [ ] Route callable factories through codec lookup before the generic function
   rejection.
 - [ ] Include callable factories in JSON encoder cycle tracking.
@@ -208,12 +208,12 @@ Expected implementation files:
 Expected implementation and test files:
 
 - `packages/data-model/src/codec-common/` for the codec/state validator
-- `packages/data-model/src/codec-json/CodecRegistry.ts`
+- `packages/data-model/src/codec-common/CodecRegistry.ts`
 - `packages/data-model/src/codec-json/JsonEncodingContext.ts`
-- `packages/data-model/src/codec-json/createDefaultRegistry.ts`
+- `packages/data-model/src/codec-common/createDefaultRegistry.ts`
 - `packages/data-model/src/codec-json/json-encoding.ts`
 - `packages/data-model/test/codec-common/FactoryCodec.test.ts`
-- `packages/data-model/test/codec-json/CodecRegistry.test.ts`
+- `packages/data-model/test/codec-common/CodecRegistry.test.ts`
 - `packages/data-model/test/codec-json/JsonEncodingContext.test.ts`
 - `packages/data-model/test/codec-json/json-encoding.test.ts`
 - a focused `packages/data-model/test/fabric-factory.test.ts`
