+++ packages/data-model/src/codec-common/createDefaultRegistry.ts
@@ -1,11 +1,11 @@
-import { CODEC } from "@/codec-common/interface.ts";
-import { BigIntCodec } from "@/codec-common/BigIntCodec.ts";
-import { SpecialNumberCodec } from "@/codec-common/SpecialNumberCodec.ts";
-import { SymbolCodec } from "@/codec-common/SymbolCodec.ts";
-import { UndefinedCodec } from "@/codec-common/UndefinedCodec.ts";
 import { codecClasses as primitiveCodecClasses } from "@/fabric-primitives/index.ts";
 import { codecClasses as instanceCodecClasses } from "@/fabric-instances/index.ts";
 
+import { CODEC } from "./interface.ts";
+import { BigIntCodec } from "./BigIntCodec.ts";
+import { SpecialNumberCodec } from "./SpecialNumberCodec.ts";
+import { SymbolCodec } from "./SymbolCodec.ts";
+import { UndefinedCodec } from "./UndefinedCodec.ts";
 import { CodecRegistry } from "./CodecRegistry.ts";
 
 /**
