+++ packages/data-model/src/codec-common/index.ts
@@ -21,3 +21,7 @@ export { UndefinedCodec } from "./UndefinedCodec.ts";
 export { BigIntCodec } from "./BigIntCodec.ts";
 export { SpecialNumberCodec } from "./SpecialNumberCodec.ts";
 export { SymbolCodec } from "./SymbolCodec.ts";
+
+// Codec registry and factory.
+export { CodecRegistry } from "./CodecRegistry.ts";
+export { createDefaultRegistry } from "./createDefaultRegistry.ts";
